requirements.txt
selenium==4.0.0
pytest==6.2.5
loguru==0.5.3
webdriver-manager==3.5.1
python-dateutil==2.7.3